+++ src/main/java/com/epam/reportportal/karate/ReportPortalHook.java
@@ -62,6 +62,11 @@ public class ReportPortalHook implements RuntimeHook {
 	private final Map<Maybe<String>, Date> stepStartTimeMap = new ConcurrentHashMap<>();
 	private volatile Thread shutDownHook;
 
+	/**
+	 * Create a new instance of the ReportPortalHook with the specified ReportPortal instance.
+	 *
+	 * @param reportPortal the ReportPortal instance
+	 */
 	public ReportPortalHook(ReportPortal reportPortal) {
 		launch = new MemoizingSupplier<>(() -> {
 			ListenerParameters params = reportPortal.getParameters();
@@ -74,10 +79,15 @@ public class ReportPortalHook implements RuntimeHook {
 		});
 	}
 
+	/**
+	 * Default constructor. Create a new instance of the ReportPortalHook with default ReportPortal instance.
+	 */
+	@SuppressWarnings("unused")
 	public ReportPortalHook() {
 		this(ReportPortal.builder().build());
 	}
 
+	@SuppressWarnings("unused")
 	public ReportPortalHook(Supplier<Launch> launchSupplier) {
 		launch = new MemoizingSupplier<>(launchSupplier);
 		shutDownHook = registerShutdownHook(this::finishLaunch);
@@ -187,7 +197,7 @@ public class ReportPortalHook implements RuntimeHook {
 	 */
 	@Nonnull
 	protected StartTestItemRQ buildStartScenarioRq(@Nonnull ScenarioRuntime sr) {
-		return ReportPortalUtils.buildStartScenarioRq(sr.scenario);
+		return ReportPortalUtils.buildStartScenarioRq(sr.result);
 	}
 
 	@Override
@@ -208,9 +218,7 @@ public class ReportPortalHook implements RuntimeHook {
 	 */
 	@Nonnull
 	protected FinishTestItemRQ buildFinishScenarioRq(@Nonnull ScenarioRuntime sr) {
-		return buildFinishTestItemRq(Calendar.getInstance().getTime(),
-				sr.result.getFailureMessageForDisplay() == null ? ItemStatus.PASSED : ItemStatus.FAILED
-		);
+		return ReportPortalUtils.buildFinishScenarioRq(sr.result);
 	}
 
 	/**
@@ -231,6 +239,7 @@ public class ReportPortalHook implements RuntimeHook {
 	 *
 	 * @param step Karate's Step object instance
 	 * @param sr   Karate's ScenarioRuntime object instance
+	 * @return item ID future
 	 */
 	public Maybe<String> startBackground(@Nonnull Step step, @Nonnull ScenarioRuntime sr) {
 		return backgroundIdMap.computeIfAbsent(sr.scenario.getUniqueId(), k -> {