+++ src/main/java/com/epam/reportportal/karate/ReportPortalPublisher.java
@@ -187,7 +187,7 @@ public class ReportPortalPublisher {
 	 */
 	@Nonnull
 	protected StartTestItemRQ buildStartScenarioRq(@Nonnull ScenarioResult scenarioResult) {
-		return ReportPortalUtils.buildStartScenarioRq(scenarioResult.getScenario());
+		return ReportPortalUtils.buildStartScenarioRq(scenarioResult);
 	}
 
 	/**
@@ -211,10 +211,7 @@ public class ReportPortalPublisher {
 	 */
 	@Nonnull
 	protected FinishTestItemRQ buildFinishScenarioRq(@Nonnull ScenarioResult scenarioResult) {
-		return buildFinishTestItemRq(Calendar.getInstance().getTime(),
-				scenarioResult.getFailureMessageForDisplay() == null ? ItemStatus.PASSED : ItemStatus.FAILED
-		);
-
+		return ReportPortalUtils.buildFinishScenarioRq(scenarioResult);
 	}
 
 	/**